add x64-Release configuration for Visual Studio.
As well as fix a minor bug in the build scripts that caused weirdness with copying the DLLs.

+++ libgame-emu-common/src/core.cpp
@@ -15,7 +15,7 @@ namespace GameEmu::Common
 	int Core::addDependency(std::string coreName)
 	{
 		dependencies.push_back(loader->getLoadedCore(coreName));
-		return dependencies.size() - 1;
+		return (int)dependencies.size() - 1;
 	}
 
 	void Core::LoadDependencies()
